fix: repository 반환타입 long으로 통일

+++ pennyway-domain/src/main/java/kr/co/pennyway/domain/domains/spending/repository/SpendingRepository.java
@@ -13,7 +13,7 @@ public interface SpendingRepository extends ExtendedRepository<Spending, Long>, 
     boolean existsByIdAndUser_Id(Long id, Long userId);
 
     @Transactional(readOnly = true)
-    Integer countByUserIdAndIdIn(Long userId, List<Long> spendingIds);
+    long countByUserIdAndIdIn(Long userId, List<Long> spendingIds);
 
     @Modifying(clearAutomatically = true)
     @Transactional
